add client only target
