+++ client/main.go
@@ -1,0 +1,137 @@
+package main
+
+import (
+	"context"
+	"fmt"
+	"log"
+	"strings"
+
+	"github.com/jmorganca/ollama/api"
+	"github.com/jmorganca/ollama/cmd"
+	"github.com/jmorganca/ollama/cmd/generate"
+	"github.com/jmorganca/ollama/cmd/listhandler"
+	"github.com/jmorganca/ollama/version"
+	"github.com/spf13/cobra"
+)
+
+func runHandler(cmd *cobra.Command, args []string) error {
+	client, err := api.ClientFromEnvironment()
+	if err != nil {
+		return err
+	}
+
+	name := args[0]
+	// check if the model exists on the server
+	_, err = client.Show(cmd.Context(), &api.ShowRequest{Name: name})
+	if err != nil {
+		return err
+	}
+
+	return generate.Run(cmd, args)
+}
+
+func checkServerHeartbeat(cmd *cobra.Command, _ []string) error {
+	client, err := api.ClientFromEnvironment()
+	if err != nil {
+		return err
+	}
+	if err := client.Heartbeat(cmd.Context()); err != nil {
+		if !strings.Contains(err.Error(), "connection refused") {
+			return err
+		}
+		return fmt.Errorf("could not connect to ollama server, check OLLAMA_HOST and OLLAMA_PORT environment variables")
+	}
+	return nil
+}
+
+func versionHandler(cmd *cobra.Command, _ []string) {
+	client, err := api.ClientFromEnvironment()
+	if err != nil {
+		return
+	}
+
+	serverVersion, err := client.Version(cmd.Context())
+	if err != nil {
+		fmt.Println("Warning: could not connect to a running Ollama instance")
+	}
+
+	if serverVersion != "" {
+		fmt.Printf("ollama version is %s\n", serverVersion)
+	}
+
+	if serverVersion != version.Version {
+		fmt.Printf("Warning: client version is %s\n", version.Version)
+	}
+}
+
+func newCLI() *cobra.Command {
+	log.SetFlags(log.LstdFlags | log.Lshortfile)
+	cobra.EnableCommandSorting = false
+
+	rootCmd := &cobra.Command{
+		Use:           "ollamac",
+		Short:         "Large language model client",
+		SilenceUsage:  true,
+		SilenceErrors: true,
+		CompletionOptions: cobra.CompletionOptions{
+			DisableDefaultCmd: true,
+		},
+		Run: func(cmd *cobra.Command, args []string) {
+			if version, _ := cmd.Flags().GetBool("version"); version {
+				versionHandler(cmd, args)
+				return
+			}
+
+			cmd.Print(cmd.UsageString())
+		},
+	}
+
+	rootCmd.Flags().BoolP("version", "v", false, "Show version information")
+
+	showCmd := &cobra.Command{
+		Use:     "show MODEL",
+		Short:   "Show information for a model",
+		Args:    cobra.ExactArgs(1),
+		PreRunE: checkServerHeartbeat,
+		RunE:    cmd.ShowHandler,
+	}
+
+	showCmd.Flags().Bool("license", false, "Show license of a model")
+	showCmd.Flags().Bool("modelfile", false, "Show Modelfile of a model")
+	showCmd.Flags().Bool("parameters", false, "Show parameters of a model")
+	showCmd.Flags().Bool("template", false, "Show template of a model")
+	showCmd.Flags().Bool("system", false, "Show system message of a model")
+
+	runCmd := &cobra.Command{
+		Use:     "run MODEL [PROMPT]",
+		Short:   "Run a model",
+		Args:    cobra.MinimumNArgs(1),
+		PreRunE: checkServerHeartbeat,
+		RunE:    runHandler,
+	}
+
+	runCmd.Flags().Bool("verbose", false, "Show timings for response")
+	runCmd.Flags().Bool("insecure", false, "Use an insecure registry")
+	runCmd.Flags().Bool("nowordwrap", false, "Don't wrap words to the next line automatically")
+	runCmd.Flags().String("format", "", "Response format (e.g. json)")
+
+	listCmd := &cobra.Command{
+		Use:     "list",
+		Aliases: []string{"ls"},
+		Short:   "List models",
+		PreRunE: checkServerHeartbeat,
+		RunE:    listhandler.Run,
+	}
+
+	rootCmd.AddCommand(
+		showCmd,
+		runCmd,
+		listCmd,
+	)
+
+	return rootCmd
+}
+
+func main() {
+	cobra.CheckErr(newCLI().ExecuteContext(context.Background()))
+}
